+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * Add alias for profile flag: mod-list
 * Convert toggle all mods button to menu items
 * Save mod list changes immediately rather than on pressing Play
+* Display mod dependencies by name when not in Modder Mode
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)